Fixed rewind on modify function
Saves of interpreter state needed to happen *when the function
is looked up*, not just on the same tick as that happening.
Before this commit state was being saved at the end of the tick if
a function body was accessed that tick.

Now we save state at the *beginning* of the tick. We might rewind
a little too far, but worst case that's into an old invocation of a
changed function. We'll probably get out of that during the tick (I'd
say certainly but other world events could change this).

This requires full world snapshots to be taken every tick (boo,
but dependable at least) instead of interpreter or even full world
snapshots on every function invocation. Using lots of small functions
should be possible, so once a tick is much better.