WIP working on missile fire on rewind
Rewinding to refire missile works the first time, but not the second.
Maybe to do with clearing the controller state?
Go back to 1 fps and example controller state to solve this.

Once this is working the next problem is that being in the ready?
callback for firing a missile doesn't seem to work. Rewinding
should work in both cases: when the events that lead to firing the
missile have not yet been set into motion, the saved events should
do this. When the script is in progress, it should be serialized
properly such that that state is resumed causes the same missile to
eventually be fired.

If this is hard enough need to redesign such that this is more testable
(or maybe just write some tests, but I doubt it's very testable now)